requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.2
python-telegram-bot==13.15
httpx>=0.25.0
anyio>=3.7.1
scikit-learn>=1.3.0
numpy>=1.24.0
regex>=2023.8.8
urllib3>=2.0.0
tqdm>=4.66.0
joblib>=1.3.0
# Дополнительные библиотеки для краулера
concurrent-log-handler>=0.9.24
# Add other ML libraries like tensorflow, or pytorch as needed 